Compress Effects API, remove GetBufferSize. (Breaks ABI!)
There's no situation where you wouldn't want the Effect data, so we return it at
CreateEffect/CloneEffect time. Similarly, you will know what the buffer size is
when you create it, so there's no need for GetBufferSize. Lastly, I had to add
SetEffectTechnique to fix a possible race condition and fix a dependency on the
MojoShader API in FNA.