Vulkan Defragmenter (#110)
Every time a resource is freed, we note that memory needs to be checked for fragmentation and start a 5-frame timer. This timer resets whenever a resource is freed to prevent unnecessary memory churning. When this timer finishes ticking down and the current frame's commands are submitted, we check to see if any allocations of memory are fragmented. If an allocation is fragmented, we mark it as deactivated and queue up GPU-to-GPU copy commands of the resources in the fragmented block. The copied-to resources are allocated through our main allocation system which allocates more memory if no required-size block is available. The copy commands are added to a special command buffer that is submitted immediately. This way the copies are performed asynchronously while the next frame is being built. However, if the user uploads buffer data, we wait for the defragmentation to finish to avoid a data race. Once the next frame is submitted, we iterate over the resources that were copied and destroy them. At this point we also check to see if any allocations are completely freed, and if so we deallocate that allocation.

The defragmentation system continues to iterate over allocations each subsequent 5 frames until it detects that memory is completely defragmented, at which point it stops until a resource is freed.

When an image view is freed, we walk the framebuffer cache and the descriptor set cache to check if that image view was used by any of those resources. If so we deactivate them to prevent problems with handle reusage.

All render targets are now allocated to dedicated memory blocks to prevent expensive copies of these resources and having to destroy and recreate a huge chain of resource dependencies.

We now mark off partitions to include alignment offset regions to avoid creating tiny free regions. The blocks are aware of their resource alignment so that copies can be performed correctly.

This commit adds an env var, FNA3D_VULKAN_DEVICE_LOCAL_HEAP_USAGE_FACTOR, which can be set to a value between 0 and 1 to artificially limit how much device local memory the application can consume.