[DONOTMERGE] Draft of low-level shader API.
Note that this implementation design is completely backwards: We're taking the
effect context and unwrapping it for every single call, when in reality we need
to add the shader calls to the FNA3D_Device and implement effects entirely in
FNA3D.c, with a single driver call to mark shaders as updated.

If you use this, you're getting blacklisted for all support! Not joking!